Renaming index.js to app.js
