+++ package.json
@@ -1,10 +1,10 @@
 {
     "name": "CukeHub",
     "version": "1.0.0",
-    "description": "CukeHub is a Node.js-based application for managing Gherkin Cucumber features, scenarios, and tags.",
-    "main": "index.js",
+    "description": "CukeHub is a feaures manager system for gherkin files based on NodeJs",
+    "main": "app.js",
     "scripts": {
-        "start": "node index.js",
+        "start": "node app.js",
         "startwd": "nodemon --exec \"npm test && node index.js\"",
         "test": "mocha"
     },
