+++ test/app.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const { server, getFiles, getScenarios } = require('../index.js');
+const { server, getFiles, getScenarios } = require('../app.js');
 const fs = require('fs');
 const path = require('path');
